requirements.txt
beautifulsoup4
speechrecognition
pyttsx3
google-cloud-texttospeech
simpleaudio
openai==0.28
pyaudio
python-dotenv~=1.1.0
selenium~=4.32.0
googlesearch-python
django~=5.2.1
pytz
PyJWT
djangorestframework~=3.16.0
django-cors-headers
psycopg2
Pillow
imagehash
channels~=4.2.2
uvicorn
django-q
asynctest~=0.13.0
daphne
google-genai
protobuf~=6.31.0
asgiref~=3.8.1
